requirements.txt
Django==3.2.6
dj_database_url==0.5.0
psycopg2==2.9.1
django-mptt==0.12.0
django-versatileimagefield==2.1
python-slugify==5.0.2